# A large speed up to `extract_fin_year`