--- a/man/extract_fin_year.Rd
+++ b/man/extract_fin_year.Rd
@@ -14,8 +14,8 @@ class. \code{\link[base:as.Date]{as.Date()}},
 can be used to store dates as an appropriate class.}
 }
 \description{
-\code{extract_fin_year} takes a date and assigns it to the
-correct financial year in the PHS specified format.
+\code{extract_fin_year} takes a date and extracts the
+correct financial year in the PHS specified format from it.
 }
 \details{
 The PHS accepted format for financial year is YYYY/YY e.g. 2017/18.