@Tina815 2023-06-20
Hi James, thanks very much for the improvement! I really like the clever way of doing this and made the method much more efficient.